move elementVisibility to text_measurements
